Add --no-old-file option parameter
When you are working with git, it can seems useless to copy old
translations in *.old files since we have access to old versions with
git. Using this setting means you do not have to remove locale/*.old
files before commit (or to add them to .gitignore). By default, *.old
files are created (this is opt-in).

Fixes typo in command given for bash_completion in README.md.